bancodehoras/requirements.txt

asgiref==3.7.2
coverage==7.2.7
Django==4.2.2
django-crispy-forms==1.14.0
django-environ==0.10.0
exceptiongroup==1.1.1
flake8==6.0.0
gunicorn==20.1.0
iniconfig==2.0.0
ldap3==2.9.1
mccabe==0.7.0
packaging==23.1
Pillow==8.2.0
pluggy==1.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycodestyle==2.10.0
pycryptodome==3.18.0
pyflakes==3.0.1
pytest==7.3.1
pytest-cov==4.1.0
pytest-django==4.5.2
pytz==2021.1
reportlab==3.5.59
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
xlwt==1.3.0
